added command line options to extraction script
